+++ Headers/state_chm_mod.F90
@@ -661,30 +661,34 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  SUBROUTINE Init_State_Chm( Input_Opt, State_Chm, State_Grid, RC )
+  SUBROUTINE Init_State_Chm( Input_Opt,       State_Chm, State_Grid,         &
+                             TaggedDiag_List, RC                            )
 !
 ! !USES:
 !
     USE ErrCode_Mod
     USE CharPak_Mod,          ONLY : To_UpperCase
-    USE CMN_Size_Mod,         ONLY : NDUST, NAER
-    USE GCKPP_Parameters,     ONLY : NSPEC
+    USE CMN_Size_Mod,         ONLY : NAER
+    USE CMN_Size_Mod,         ONLY : NDUST
+    USE GcKpp_Parameters,     ONLY : NSPEC
     USE Input_Opt_Mod,        ONLY : OptInput
     USE Species_Database_Mod, ONLY : Init_Species_Database
     USE State_Grid_Mod,       ONLY : GrdState
+    USE TaggedDiagList_Mod,   ONLY : TaggedDgnList
 !
 ! !INPUT PARAMETERS:
 !
-    TYPE(GrdState), INTENT(IN)    :: State_Grid  ! Grid State object
+    TYPE(GrdState),      INTENT(IN)    :: State_Grid      ! Grid State object
+    TYPE(TaggedDgnList), INTENT(IN)    :: TaggedDiag_List ! Tagged Diag List
 !
 ! !INPUT/OUTPUT PARAMETERS:
 !
-    TYPE(OptInput), INTENT(INOUT) :: Input_Opt   ! Input Options object
-    TYPE(ChmState), INTENT(INOUT) :: State_Chm   ! Chemistry State object
+    TYPE(OptInput),      INTENT(INOUT) :: Input_Opt       ! Input Options
+    TYPE(ChmState),      INTENT(INOUT) :: State_Chm       ! Chemistry State
 !
 ! !OUTPUT PARAMETERS:
 !
-    INTEGER,        INTENT(OUT)   :: RC          ! Return code
+    INTEGER,             INTENT(OUT)   :: RC              ! Return code
 !
 ! !REMARKS:
 !  In the near future we will put some error trapping on the allocations
@@ -933,7 +937,7 @@ CONTAINS
     !========================================================================
     ! Initialize the 1-D mapping vectors (e.g. State_Chm%Map_DryDep)
     !========================================================================
-    CALL Init_Mapping_Vectors( Input_Opt, State_Chm, RC )
+    CALL Init_Mapping_Vectors( Input_Opt, State_Chm, TaggedDiag_List, RC )
     IF ( RC /= GC_SUCCESS ) THEN
        errMsg = 'Error encountered in "Init_Mapping_Vectors" routine!'
        CALL GC_Error( errMsg, RC, thisLoc )
@@ -2379,21 +2383,26 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  SUBROUTINE Init_Mapping_Vectors( Input_Opt, State_Chm, RC )
+  SUBROUTINE Init_Mapping_Vectors( Input_Opt, State_Chm, TaggedDiag_List, RC )
 !
 ! !USES:
 !
-    USE GCKPP_Parameters, ONLY : NSPEC
-    USE Input_Opt_Mod,    ONLY : OptInput
+    USE GcKpp_Parameters,   ONLY : NSPEC
+    USE Input_Opt_Mod,      ONLY : OptInput
+    USE TaggedDiagList_Mod, ONLY : TaggedDgnList
 !
 ! !INPUT/OUTPUT PARAMETERS:
 !
-    TYPE(OptInput), INTENT(INOUT) :: Input_Opt   ! Input Options object
-    TYPE(ChmState), INTENT(INOUT) :: State_Chm   ! Chemistry State object
+    TYPE(TaggedDgnList), INTENT(IN)    :: TaggedDiag_List ! Tagged Diag List
+!
+! !INPUT/OUTPUT PARAMETERS:
+!
+    TYPE(OptInput),      INTENT(INOUT) :: Input_Opt       ! Input Options
+    TYPE(ChmState),      INTENT(INOUT) :: State_Chm       ! Chemistry State
 !
 ! !OUTPUT PARAMETERS:
 !
-    INTEGER,        INTENT(OUT)   :: RC          ! Success or failure
+    INTEGER,             INTENT(OUT)   :: RC              ! Success or failure
 !
 ! !REVISION HISTORY:
 !  06 Jan 2015 - R. Yantosca - Initial version
@@ -2752,7 +2761,7 @@ CONTAINS
     ! Set up the mapping for PRODUCTION AND LOSS DIAGNOSTIC SPECIES
     !------------------------------------------------------------------------
     IF ( State_Chm%nProd > 0 .or. State_Chm%nLoss > 0 ) THEN
-       CALL MapProdLossSpecies( Input_Opt, State_Chm, RC )
+       CALL MapProdLossSpecies( Input_Opt, State_Chm, TaggedDiag_List, RC )
        IF ( RC /= GC_SUCCESS ) THEN
           ErrMsg = 'Error encountered in routine "MapProdLossSpecies"!'
           CALL GC_Error( ErrMsg, RC, ThisLoc )
@@ -6942,22 +6951,31 @@ CONTAINS
 !\\
 ! !INTERFACE:
 !
-  SUBROUTINE MapProdLossSpecies( Input_Opt, State_Chm, RC )
+  SUBROUTINE MapProdLossSpecies( Input_Opt, State_Chm, TaggedDiag_List, RC )
 !
 ! !USES:
 !
-    USE GcKpp_Monitor,    ONLY : Fam_Names
-    USE GcKpp_Parameters, ONLY : nFam
-    USE Input_Opt_Mod,    ONLY : OptInput
+    USE CharPak_Mod,        ONLY : To_UpperCase
+    USE GcKpp_Monitor,      ONLY : Fam_Names
+    USE GcKpp_Parameters,   ONLY : nFam
+    USE Input_Opt_Mod,      ONLY : OptInput
+    USE TaggedDiagList_Mod, ONLY : DgnTagList
+    USE TaggedDiagList_Mod, ONLY : TaggedDgnList
+    USE TaggedDiagList_Mod, ONLY : Query_Tag_in_TagList
+    USE TaggedDiagList_Mod, ONLY : Query_TaggedDiagList
 !
 ! !INPUT/OUTPUT PARAMETERS:
 !
-    TYPE(OptInput), INTENT(INOUT) :: Input_Opt   ! Input Options object
-    TYPE(ChmState), INTENT(INOUT) :: State_Chm   ! Chemistry State object
+    TYPE(TaggedDgnList), INTENT(IN)    :: TaggedDiag_List ! Tag Diag. List
+!
+! !INPUT/OUTPUT PARAMETERS:
+!
+    TYPE(OptInput),      INTENT(INOUT) :: Input_Opt       ! Input Options
+    TYPE(ChmState),      INTENT(INOUT) :: State_Chm       ! Chemistry State
 !
 ! !OUTPUT PARAMETERS:
 !
-    INTEGER,        INTENT(OUT)   :: RC          ! Return code
+    INTEGER,              INTENT(OUT)  :: RC              ! Return code
 !
 ! !REMARKS:
 !
@@ -6971,23 +6989,29 @@ CONTAINS
 ! !LOCAL VARIABLES:
 !
     ! Scalars
-    INTEGER            :: Id,     N
-    INTEGER            :: P,      L
+    LOGICAL                   :: found
+    INTEGER                   :: Id, N, numTags, numWildCards, P, L
 
     ! Strings
-    CHARACTER(LEN=36)  :: Name
-    CHARACTER(LEN=255) :: ErrMsg, ThisLoc
+    CHARACTER(LEN=36)         :: name
+    CHARACTER(LEN=255)        :: diagName
+    CHARACTER(LEN=255)        :: errMsg
+    CHARACTER(LEN=255)        :: thisLoc
+
+    ! Objects
+    TYPE(DgnTagList)          :: TagList
+    TYPE(DgnTagList)          :: WildCardList
 
     !=======================================================================
     ! GetProdLossSpecies begins here!
     !=======================================================================
 
     ! Initialize
-    RC      = GC_SUCCESS
-    P       = 0
-    L       = 0
-    ErrMsg  = ''
-    ThisLoc = &
+    RC       =  GC_SUCCESS
+    P        =  0
+    L        =  0
+    ErrMsg   =  ''
+    ThisLoc  =  &
          ' -> at MapProdLossSpecies (in module Headers/state_chm_mod.F90)'
 
     !=======================================================================
@@ -7002,28 +7026,104 @@ CONTAINS
        ! Loop over the number of prod/loss species
        DO N = 1, nFam
 
+          ! Initialize
+          name     = ''
+          diagName = ''
+          found    = .FALSE.
+
           ! Get the KPP prod/loss species from the FAM_NAMES
           ! array in the gckpp_Parameters.F90 module.
           ! NOTE: This is the KPP ID number (index of "VAR" array)
           ! and not the GEOS-Chem "main" species index!!!
           Id = Ind_( TRIM( Fam_Names(N) ), 'K' )
 
-          ! Add the species
+          ! Try to add the species if it is valid
           IF ( Id > 0 ) THEN
 
              ! KPP prod/loss species name
              Name = TRIM( Fam_Names(N) )
 
-             ! Fix the name so that it is of the form Prod_<spcname> or
-             ! Loss_<spcname>.  This will facilitate the new diagnostics.
+             !---------------------------------------------------------------
+             ! Loss species
+             !---------------------------------------------------------------
              IF ( Name(1:1) == 'L' ) THEN
-                L                      = L + 1
-                State_Chm%Map_Loss(L)  = Id
-                State_Chm%Name_Loss(L) = 'Loss_' // TRIM( Name(2:) )
+
+                ! The diagnostic name will be `Loss_' + the KPP family name
+                diagName = 'Loss_' // TRIM( Name(2:) )
+
+                ! Check if the TaggedDiag_List exists for "Loss"
+                CALL Query_TaggedDiagList(                                   &
+                     TaggedDiagList = TaggedDiag_List,                       &
+                     diagName       = 'Loss',                                &
+                     found          = found,                                 &
+                     numTags        = numTags,                               &
+                     TagList        = TagList,                               &
+                     numWildCards   = numWildCards,                          &
+                     WildCardList   = WildCardList,                          &
+                     RC             = RC                                    )
+
+                ! Check if the given loss species or wildcard has been
+                ! requested for diagnostic archival before updating
+                ! mapping vectors.
+                IF ( found ) THEN
+
+                   IF ( numTags > 0 ) THEN
+                      CALL Query_Tag_in_TagList( TagList,       name(2:),    &
+                                                 found,         RC          )
+                   ELSE IF ( numWildCards > 0 ) THEN
+                      CALL Query_Tag_in_TagList( WildCardList, 'LOS',        &
+                                                 found,         RC          )
+                   ENDIF
+
+                   IF ( found ) THEN
+                      L                      = L + 1
+                      State_Chm%Map_Loss(L)  = Id
+                      State_Chm%Name_Loss(L) = TRIM( diagName )
+                   ENDIF
+                ENDIF
+
+             !---------------------------------------------------------------
+             ! Prod species
+             !---------------------------------------------------------------
              ELSE IF ( Name(1:1) == 'P' ) THEN
-                P                      = P + 1
-                State_Chm%Map_Prod(P)  = Id
-                State_Chm%Name_Prod(P) = 'Prod_' // TRIM( Name(2:) )
+
+                ! The diagnostic name will be `Prod_' + the KPP family name
+                diagName = 'Prod_' // TRIM( Name(2:) )
+
+                ! Check if the TaggedDiag_List exists for "Prod"
+                CALL Query_TaggedDiagList(                                   &
+                     TaggedDiagList = TaggedDiag_List,                       &
+                     diagName       = 'Prod',                                &
+                     found          = found,                                 &
+                     numTags        = numTags,                               &
+                     TagList        = TagList,                               &
+                     numWildCards   = numWildCards,                          &
+                     WildCardList   = WildCardList,                          &
+                     RC             = RC                                    )
+
+                ! Check if the given prod species or wildcard has been
+                ! requested for diagnostic archival before updating
+                ! mapping vectors.
+                IF ( found ) THEN
+
+                   IF ( numTags > 0 ) THEN
+                      CALL Query_Tag_in_TagList( TagList,       name(2:),    &
+                                                 found,         RC          )
+                   ELSE IF ( numWildCards > 0 ) THEN
+                      CALL Query_Tag_in_TagList( WildCardList, 'PRD',        &
+                                                 found,         RC          )
+                   ENDIF
+
+                   IF ( found ) THEN
+                      P                      = P + 1
+                      State_Chm%Map_Prod(P)  = Id
+                      State_Chm%Name_Prod(P) = TRIM( diagName )
+                   ENDIF
+                ENDIF
+
+             !---------------------------------------------------------------
+             ! Invalid prod/loss species
+             !---------------------------------------------------------------
              ELSE
                 ErrMsg = 'Invalid prod/loss species name!' //                &
                           TRIM( Fam_Names(N) )
@@ -7031,6 +7131,9 @@ CONTAINS
                 RETURN
              ENDIF
 
+          !------------------------------------------------------------------
+          ! Could not find KPP species
+          !------------------------------------------------------------------
           ELSE
 
              ! Invalid species, exit with error!